[#175890863] Adds navigation from methods list and replace actions
