+++ ts/features/wallet/onboarding/bancomatPay/screens/search/BPaySearchStartScreen.tsx
@@ -1,72 +1,57 @@
-import { Content } from "native-base";
 import * as React from "react";
-import { SafeAreaView } from "react-native";
 import { connect } from "react-redux";
 import { Dispatch } from "redux";
-import { IOStyles } from "../../../../../../components/core/variables/IOStyles";
 import { withLightModalContext } from "../../../../../../components/helpers/withLightModalContext";
-import BaseScreenComponent from "../../../../../../components/screens/BaseScreenComponent";
-import FooterWithButtons from "../../../../../../components/ui/FooterWithButtons";
 import { LightModalContextInterface } from "../../../../../../components/ui/LightModal";
-import I18n from "../../../../../../i18n";
 import { GlobalState } from "../../../../../../store/reducers/types";
-import { emptyContextualHelp } from "../../../../../../utils/emptyContextualHelp";
-import {
-  cancelButtonProps,
-  confirmButtonProps
-} from "../../../../../bonus/bonusVacanze/components/buttons/ButtonConfigurations";
+import TosBonusComponent from "../../../../../bonus/bonusVacanze/components/TosBonusComponent";
 import { isError, isLoading } from "../../../../../bonus/bpd/model/RemoteValue";
-import { loadAbi } from "../../../bancomat/store/actions";
 import { abiSelector } from "../../../store/abi";
-import {
-  navigateToOnboardingBPayChooseBank,
-  navigateToOnboardingBPaySearchAvailableUserAccount
-} from "../../navigation/action";
+import SearchStartScreen from "../../../common/searchBank/SearchStartScreen";
 import {
   searchUserBPay,
   walletAddBPayBack,
   walletAddBPayCancel
 } from "../../store/actions";
-import SectionStatusComponent from "../../../../../../components/SectionStatusComponent";
+import {
+  navigateToOnboardingBPayChooseBank,
+  navigateToOnboardingBPaySearchAvailableUserAccount
+} from "../../navigation/action";
 
 type Props = LightModalContextInterface &
   ReturnType<typeof mapStateToProps> &
   ReturnType<typeof mapDispatchToProps>;
 
-const renderFooterButtons = (onCancel: () => void, onContinue: () => void) => (
-  <FooterWithButtons
-    type={"TwoButtonsInlineThird"}
-    leftButton={cancelButtonProps(onCancel, I18n.t("global.buttons.cancel"))}
-    rightButton={confirmButtonProps(
-      onContinue,
-      I18n.t("global.buttons.continue")
-    )}
-  />
-);
+const tos_url = "https://io.italia.it/app-content/privacy_bancomat.html";
 
 /**
- * This screen allows the user to choose a specific bank to search for their BPay accounts.
+ * This screen allows the user to choose a specific bank to search for their Bancomat.
  * @constructor
  */
-const BPaySearchStartScreen = (props: Props): React.ReactElement => (
-  <BaseScreenComponent
-    goBack={true}
-    headerTitle={I18n.t("wallet.searchAbi.header")}
-    contextualHelp={emptyContextualHelp}
-  >
-    <SafeAreaView style={IOStyles.flex}>
-      <Content style={IOStyles.flex} />
-      <SectionStatusComponent sectionKey={"bancomatpay"} />
-      {renderFooterButtons(props.onCancel, props.searchBPay)}
-    </SafeAreaView>
-  </BaseScreenComponent>
-);
+const BPaySearchStartScreen: React.FunctionComponent<Props> = (
+  props: Props
+) => {
+  const openTosModal = () => {
+    props.showModal(
+      <TosBonusComponent tos_url={tos_url} onClose={props.hideModal} />
+    );
+  };
+
+  return (
+    <SearchStartScreen
+      methodName={"BANCOMAT Pay"}
+      onCancel={props.onCancel}
+      onSearch={props.searchAccounts}
+      handleTosModal={openTosModal}
+      navigateToSearchBank={props.navigateToSearchBankScreen}
+    />
+  );
+};
 
 const mapDispatchToProps = (dispatch: Dispatch) => ({
-  loadAbis: () => dispatch(loadAbi.request()),
   onCancel: () => dispatch(walletAddBPayCancel()),
   onBack: () => dispatch(walletAddBPayBack()),
-  searchBPay: (abi?: string) => {
+  searchAccounts: (abi?: string) => {
     dispatch(searchUserBPay.request(abi));
     dispatch(navigateToOnboardingBPaySearchAvailableUserAccount());
   },
